E2E tests: remove dead code
