+++ test/automation/src/index.ts
@@ -3,29 +3,23 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-export * from './activityBar';
+// --- Start Positron ---
+// Removed existing exports
+// --- End Positron ---
 export * from './application';
 export * from './code';
-export * from './debug';
-export * from './editor';
-export * from './editors';
-export * from './explorer';
-export * from './extensions';
-export * from './keybindings';
+// --- Start Positron ---
+// Removed existing exports
+// --- End Positron ---
 export * from './logger';
-export * from './peek';
-export * from './problems';
-export * from './quickinput';
-export * from './quickaccess';
-export * from './scm';
-export * from './search';
-export * from './settings';
-export * from './statusbar';
-export * from './terminal';
-export * from './viewlet';
-export * from './localization';
+// --- Start Positron ---
+// Removed existing exports
+// --- End Positron ---
 export * from './workbench';
-export * from './task';
+// --- Start Positron ---
+// Removed existing exports
+// --- End Positron ---
+
 // --- Start Positron ---
 export * from './positron/positronConsole';
 export * from './positron/positronPopups';
