+++ test/automation/src/workbench.ts
@@ -3,25 +3,13 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { Explorer } from './explorer';
-import { ActivityBar } from './activityBar';
-import { QuickAccess } from './quickaccess';
-import { QuickInput } from './quickinput';
-import { Extensions } from './extensions';
-import { Search } from './search';
-import { Editor } from './editor';
-import { SCM } from './scm';
-import { Debug } from './debug';
-import { StatusBar } from './statusbar';
-import { Problems } from './problems';
-import { SettingsEditor } from './settings';
-import { KeybindingsEditor } from './keybindings';
-import { Editors } from './editors';
+// --- Start Positron ---
+// Removed existing imports
+// --- End Positron ---
 import { Code } from './code';
-import { Terminal } from './terminal';
-import { Notebook } from './notebook';
-import { Localization } from './localization';
-import { Task } from './task';
+// --- Start Positron ---
+// Removed existing imports
+// --- End Positron ---
 
 // --- Start Positron ---
 import { PositronInterpreterDropdown } from './positron/positronInterpreterDropdown';
@@ -59,26 +47,10 @@ export interface Commands {
 
 export class Workbench {
 
-	readonly quickaccess: QuickAccess;
-	readonly quickinput: QuickInput;
-	readonly editors: Editors;
-	readonly explorer: Explorer;
-	readonly activitybar: ActivityBar;
-	readonly search: Search;
-	readonly extensions: Extensions;
-	readonly editor: Editor;
-	readonly scm: SCM;
-	readonly debug: Debug;
-	readonly statusbar: StatusBar;
-	readonly problems: Problems;
-	readonly settingsEditor: SettingsEditor;
-	readonly keybindingsEditor: KeybindingsEditor;
-	readonly terminal: Terminal;
-	readonly notebook: Notebook;
-	readonly localization: Localization;
-	readonly task: Task;
-
 	// --- Start Positron ---
+
+	// removed existing properties
+
 	readonly positronInterpreterDropdown: PositronInterpreterDropdown;
 	readonly positronPopups: PositronPopups;
 	readonly positronConsole: PositronConsole;
@@ -109,26 +81,10 @@ export class Workbench {
 	// --- End Positron ---
 
 	constructor(code: Code) {
-		this.editors = new Editors(code);
-		this.quickinput = new QuickInput(code);
-		this.quickaccess = new QuickAccess(code, this.editors, this.quickinput);
-		this.explorer = new Explorer(code);
-		this.activitybar = new ActivityBar(code);
-		this.search = new Search(code);
-		this.extensions = new Extensions(code, this.quickaccess);
-		this.editor = new Editor(code, this.quickaccess);
-		this.scm = new SCM(code);
-		this.debug = new Debug(code, this.quickaccess, this.editors, this.editor);
-		this.statusbar = new StatusBar(code);
-		this.problems = new Problems(code, this.quickaccess);
-		this.settingsEditor = new SettingsEditor(code, this.editors, this.editor, this.quickaccess);
-		this.keybindingsEditor = new KeybindingsEditor(code);
-		this.terminal = new Terminal(code, this.quickaccess, this.quickinput);
-		this.notebook = new Notebook(this.quickaccess, this.quickinput, code);
-		this.localization = new Localization(code);
-		this.task = new Task(code, this.editor, this.editors, this.quickaccess, this.quickinput, this.terminal);
-
 		// --- Start Positron ---
+
+		// removed existing initializations
+
 		this.positronPopups = new PositronPopups(code);
 		this.positronInterpreterDropdown = new PositronInterpreterDropdown(code);
 		this.positronVariables = new PositronVariables(code);
