--- test/smoke/.yarnrc
@@ -1,5 +1,0 @@
-disturl ""
-target ""
-ms_build_id ""
-runtime "node"
-build_from_source ""
