--- test/smoke/README.md
@@ -1,6 +1,0 @@
-<!-- Start Positron -->
-<!-- If you are seeking the original Vscode README go here: https://github.com/microsoft/vscode/tree/main/test/smoke -->
-
-These tests are currently not active. Please see the [E2E tests section](https://github.com/posit-dev/positron/blob/main/test/e2e/README.md).
-
-<!-- End Positron -->
