--- test/smoke/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "code-oss-dev-smoke-test",
-  "version": "0.1.0",
-  "license": "MIT",
-  "main": "./src/main.js",
-  "scripts": {
-    "compile": "cd ../automation && npm run compile && cd ../smoke && node ../../node_modules/typescript/bin/tsc",
-    "watch-automation": "cd ../automation && npm run watch",
-    "watch-smoke": "node ../../node_modules/typescript/bin/tsc --watch --preserveWatchOutput",
-    "watch": "npm-run-all -lp watch-automation watch-smoke",
-    "mocha": "node ../node_modules/mocha/bin/mocha"
-  },
-  "dependencies": {
-    "canvas": "^2.11.2",
-    "mkdirp": "^1.0.4",
-    "ncp": "^2.0.0",
-    "node-fetch": "^2.6.7",
-    "resemblejs": "5.0.0",
-    "rimraf": "3.0.2"
-  },
-  "devDependencies": {
-    "@types/mocha": "^9.1.1",
-    "@types/ncp": "2.0.1",
-    "@types/node": "20.x",
-    "@types/node-fetch": "^2.5.10",
-    "@types/resemblejs": "4.1.3",
-    "@types/rimraf": "3.0.2",
-    "archiver": "^7.0.1",
-    "npm-run-all": "^4.1.5",
-    "watch": "^1.0.2"
-  }
-}
