--- test/smoke/src/areas/terminal/terminal-persistence.test.ts
@@ -1,83 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { Application, Terminal, TerminalCommandId, TerminalCommandIdWithValue, SettingsEditor } from '../../../../automation';
-import { setTerminalTestSettings } from './terminal-helpers';
-
-export function setup(options?: { skipSuite: boolean }) {
-	(options?.skipSuite ? describe.skip : describe)('Terminal Persistence', () => {
-		// Acquire automation API
-		let terminal: Terminal;
-		let settingsEditor: SettingsEditor;
-
-		before(async function () {
-			const app = this.app as Application;
-			terminal = app.workbench.terminal;
-			settingsEditor = app.workbench.settingsEditor;
-			await setTerminalTestSettings(app);
-		});
-
-		after(async function () {
-			await settingsEditor.clearUserSettings();
-		});
-
-		describe('detach/attach', () => {
-			// https://github.com/microsoft/vscode/issues/137799
-			it('should support basic reconnection', async () => {
-				await terminal.createTerminal();
-				// TODO: Handle passing in an actual regex, not string
-				await terminal.assertTerminalGroups([
-					[{ name: '.*' }]
-				]);
-
-				// Get the terminal name
-				await terminal.assertTerminalGroups([
-					[{ name: '.*' }]
-				]);
-				const name = (await terminal.getTerminalGroups())[0][0].name!;
-
-				// Detach
-				await terminal.runCommand(TerminalCommandId.DetachSession);
-				await terminal.assertTerminalViewHidden();
-
-				// Attach
-				await terminal.runCommandWithValue(TerminalCommandIdWithValue.AttachToSession, name);
-				await terminal.assertTerminalGroups([
-					[{ name }]
-				]);
-			});
-
-			it.skip('should persist buffer content', async () => {
-				await terminal.createTerminal();
-				// TODO: Handle passing in an actual regex, not string
-				await terminal.assertTerminalGroups([
-					[{ name: '.*' }]
-				]);
-
-				// Get the terminal name
-				await terminal.assertTerminalGroups([
-					[{ name: '.*' }]
-				]);
-				const name = (await terminal.getTerminalGroups())[0][0].name!;
-
-				// Write in terminal
-				await terminal.runCommandInTerminal('echo terminal_test_content');
-				await terminal.waitForTerminalText(buffer => buffer.some(e => e.includes('terminal_test_content')));
-
-				// Detach
-				await terminal.runCommand(TerminalCommandId.DetachSession);
-				await terminal.assertTerminalViewHidden();
-
-				// Attach
-				await terminal.runCommandWithValue(TerminalCommandIdWithValue.AttachToSession, name);
-				await terminal.assertTerminalGroups([
-					[{ name }]
-				]);
-				// There can be line wrapping, so remove newlines and carriage returns #216464
-				await terminal.waitForTerminalText(buffer => buffer.some(e => e.replaceAll(/[\r\n]/g, '').includes('terminal_test_content')));
-			});
-		});
-	});
-}
