--- test/smoke/src/areas/terminal/terminal-tabs.test.ts
@@ -1,110 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { Application, Terminal, TerminalCommandId, TerminalCommandIdWithValue, SettingsEditor } from '../../../../automation';
-import { setTerminalTestSettings } from './terminal-helpers';
-
-export function setup(options?: { skipSuite: boolean }) {
-	(options?.skipSuite ? describe.skip : describe)('Terminal Tabs', () => {
-		// Acquire automation API
-		let terminal: Terminal;
-		let settingsEditor: SettingsEditor;
-
-		before(async function () {
-			const app = this.app as Application;
-			terminal = app.workbench.terminal;
-			settingsEditor = app.workbench.settingsEditor;
-			await setTerminalTestSettings(app);
-		});
-
-		after(async function () {
-			await settingsEditor.clearUserSettings();
-		});
-
-		it('clicking the plus button should create a terminal and display the tabs view showing no split decorations', async () => {
-			await terminal.createTerminal();
-			await terminal.clickPlusButton();
-			await terminal.assertTerminalGroups([[{}], [{}]]);
-		});
-
-		it('should rename the single tab', async () => {
-			await terminal.createTerminal();
-			const name = 'my terminal name';
-			await terminal.runCommandWithValue(TerminalCommandIdWithValue.Rename, name);
-			await terminal.assertSingleTab({ name });
-		});
-
-		// DEBT: Flaky https://github.com/microsoft/vscode/issues/216564
-		it.skip('should reset the tab name to the default value when no name is provided', async () => {
-			await terminal.createTerminal();
-			const defaultName = await terminal.getSingleTabName();
-			const name = 'my terminal name';
-			await terminal.runCommandWithValue(TerminalCommandIdWithValue.Rename, name);
-			await terminal.assertSingleTab({ name });
-			await terminal.runCommandWithValue(TerminalCommandIdWithValue.Rename, undefined);
-			await terminal.assertSingleTab({ name: defaultName });
-		});
-
-		it('should rename the tab in the tabs list', async () => {
-			await terminal.createTerminal();
-			await terminal.runCommand(TerminalCommandId.Split);
-			const name = 'my terminal name';
-			await terminal.runCommandWithValue(TerminalCommandIdWithValue.Rename, name);
-			await terminal.assertTerminalGroups([[{}, { name }]]);
-		});
-
-		it('should create a split terminal when single tab is alt clicked', async () => {
-			await terminal.createTerminal();
-			const page = await terminal.getPage();
-			page.keyboard.down('Alt');
-			await terminal.clickSingleTab();
-			page.keyboard.up('Alt');
-			await terminal.assertTerminalGroups([[{}, {}]]);
-		});
-
-		it('should do nothing when join tabs is run with only one terminal', async () => {
-			await terminal.runCommand(TerminalCommandId.Show);
-			await terminal.runCommand(TerminalCommandId.Join);
-			await terminal.assertTerminalGroups([[{}]]);
-		});
-
-		it('should do nothing when join tabs is run with only split terminals', async () => {
-			await terminal.runCommand(TerminalCommandId.Show);
-			await terminal.runCommand(TerminalCommandId.Split);
-			await terminal.runCommand(TerminalCommandId.Join);
-			await terminal.assertTerminalGroups([[{}], [{}]]);
-		});
-
-		it('should join tabs when more than one non-split terminal', async () => {
-			await terminal.runCommand(TerminalCommandId.Show);
-			await terminal.createTerminal();
-			await terminal.runCommand(TerminalCommandId.Join);
-			await terminal.assertTerminalGroups([[{}, {}]]);
-		});
-
-		it('should do nothing when unsplit tabs called with no splits', async () => {
-			await terminal.runCommand(TerminalCommandId.Show);
-			await terminal.createTerminal();
-			await terminal.assertTerminalGroups([[{}], [{}]]);
-			await terminal.runCommand(TerminalCommandId.Unsplit);
-			await terminal.assertTerminalGroups([[{}], [{}]]);
-		});
-
-		it('should unsplit tabs', async () => {
-			await terminal.runCommand(TerminalCommandId.Show);
-			await terminal.runCommand(TerminalCommandId.Split);
-			await terminal.assertTerminalGroups([[{}, {}]]);
-			await terminal.runCommand(TerminalCommandId.Unsplit);
-			await terminal.assertTerminalGroups([[{}], [{}]]);
-		});
-
-		it('should move the terminal to the editor area', async () => {
-			await terminal.runCommand(TerminalCommandId.Show);
-			await terminal.assertSingleTab({});
-			await terminal.runCommand(TerminalCommandId.MoveToEditor);
-			await terminal.assertEditorGroupCount(1);
-		});
-	});
-}
